dbusutils: Fix signature
The `int rowStride` is part of the DBus contract, this must not be changed.
Instead, cast it to `int` explicitly.

+++ host/abstractkrunnerplugin.cpp
@@ -63,7 +63,7 @@ RemoteImage AbstractKRunnerPlugin::serializeImage(const QImage &image)
     RemoteImage remoteImage{
         convertedImage.width(),
         convertedImage.height(),
-        convertedImage.bytesPerLine(),
+        static_cast<int>(convertedImage.bytesPerLine()),
         true, // hasAlpha
         8, // bitsPerSample
         4, // channels
